// quantri_client/src/pages/dvc/thuphilephi/models/YeuCauThanhToan.tsx
import { IBaseExt } from "@/models";

export interface IYeuCauThanhToan extends IBaseExt {
  hoSoId?: string;
  maHoSo?: string;
  ma?: string;
  soTien?: string;
  phi: string;
  lePhi: string;
  lePhiTheoTTHC?: string;
  phiTheoTTHC?: string;
  tenTTHC?: string;
  trangThai?: string;
  ngayYeuCau?: string;
  nguoiYeuCau?: string;
  tenNguoiYeuCau?: string;
  taiKhoanNguoiYeuCau?: string;
  donViThu?: string;
  hinhThucThanhToan?: string;
  hinhThucThu?: string;
  chiTiet?: string;
  ghiChuThanhToan?: string;
  mauSoBienLai?: string;
  kyHieuBienLai?: string;
  soHieuBienLai?: string;
  nguoiThuPhi?: string;
  ngayThuPhi?: string;
  ngayTiepNhan?: string;
  donViThuPhiMaSoThue?: string;
  donViMaSoThue?: string;
  ngayHoanPhi?: string;
  nguoiHoanPhi?: string;
  lyDoHoanPhi?: string;
  ngayHuy?: string;
  nguoiHuy?: string;
  lyDoHuy?: string;
  ghiChu?: string;
  tenDonVi?: string;
  tenDonViThu?: string;
  chuHoSo?: string;
  diaChiChuHoSo?: string;
  trangThaiHoSoId?: string;
  nguoiUyQuyen?: string;
  diaChiNguoiUyQuyen?: string;
  trichYeuHoSo?: string;
  nguoiNopTienBienLai?: string;
  maSoThueBienLai?: string;
  diaChiBienLai?: string;
  noiDung?: string;
  tenPhiBienLai?: string;
  tenLePhiBienLai?: string;
  kenhThucHien?: string;
  tenNganHangHoanPhi?: string;
  soTaiKhoanHoanPhi?: string;
  tenTaiKhoanHoanPhi?: string;
  soBienLai?: string;
  loaiBienLaiThanhToan?: string
  soBienLaiPhi?: string;
  soBienLaiLePhi?: string;
  soDienThoaiNguoiNopTienBienLai?: string;
  maThamChieuGiaoDich?: string;
}